requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
google-generativeai==0.3.2
python-multipart==0.0.6
pdf2image==1.17.0
Pillow==10.2.0
pydantic==2.6.0
